+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Added
 
-- `slo report` now sends the generated report to Reliably
+- `slo report` sends the generated report to Reliably
+- `slo report` shows the current SLO value progression, compared to the previous report; [#255](https://github.com/reliablyhq/cli/issues/255)
+- `slo report` shows the SLO trend for the last 5 reports; [#255](https://github.com/reliablyhq/cli/issues/255)
 
 ### Changed